+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## et-orbi 1.1.3  not yet released
 
+- Introduce EtOrbi::EoTime#ambiguous?
 - Introduce EtOrbi::EoTime#to_z for precise timezones (not offsets)